backend/cv modifier/requirements_ocr.txt
# Additional requirements for OCR-based PDF masking
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
